Add CreateReadwriteSplittingRuleStatementUpdater
