+++ shardingsphere-proxy/shardingsphere-proxy-backend/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.distsql.RDLUpdater
@@ -18,6 +18,8 @@
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterShardingTableRuleStatementUpdater
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterShardingBindingTableRuleStatementUpdater
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterShardingBroadcastTableRuleStatementUpdater
+
+org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.CreateReadwriteSplittingRuleStatementUpdater
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.AlterReadwriteSplittingRuleStatementUpdater
 
 org.apache.shardingsphere.proxy.backend.text.distsql.rdl.impl.updater.CreateDatabaseDiscoveryRuleStatementUpdater
